requirements.txt
celery==5.2.3
Flask==2.0.2
Flask-Admin==1.6.0
marshmallow==3.14.1
peewee==3.14.8
psycopg2==2.9.3
python-telegram-bot==13.11
redis==4.1.4
wtf-peewee==3.0.3
